CHROMIUM: iwl7000: chromeOS: remove erroneous cfg80211 version ifdefs
Kirtika reported this problem, and the reason is that the
backport from our master branch to the release branches
was including the new ifdefs that were only correct on
the master branch due to other changes.

This resulted in a WARN_ON() and the interface not being
able to be brought up.

(or rather, the backport thereof)

Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: ee96483107252c50e3c5c2611b7bdf7b12ed6f54