CHROMIUM: media/s5p-jpeg: fix recursive spinlock acquisition
v4l2_m2m_job_finish(), which is called from the interrupt handler with
slock acquired, can call the device_run() hook immediately if another
context was in the queue. On the s5p driver, this hook also acquires
slock, resulting in a deadlock for this particular scenario.

Fix this by releasing slock in the interrupt handler before calling
v4l2_m2m_job_finish(). This is safe to do as the state of the hardware
cannot change before v4l2_m2m_job_finish() is called anyway.

Ideally v4l2_m2m_job_finish() should be called from a bottom half
interrupt handler as it does a lot of things, but this quick fix is
helpful in mitigating the issue.

BUG=chromium:705971
TEST=video_JpegDecodeAccelerator now passes successfully

Change-Id: Ic6358c1e00fed0ea0a61f84492ceb3d594df0a4a
Signed-off-by: Alexandre Courbot <acourbot@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/481579
Reviewed-by: Tomasz Figa <tfiga@chromium.org>