CHROMIUM: iwl7000: xvt: don't sleep in Rx path
All the Rx path of the operation mode is run under a
spinlock held by the transport layer (rxq_lock). This
means that the operation is not allowed to sleep in
the Rx path.
Replace the mutex by a spin_lock to avoid this.

ticket=none

Signed-off-by: Moria Zuberi <moriax.zuberi@intel.com>
iwl7000-tree: bf49068949cb2f18131f1c72b217947eeb09cd52