CHROMIUM: iwl7000: introduce Image Loader (IML) - new firmware image
In 22650 devices a new image is introduced - IML. The IML,
image loader, is loaded by the ROM, and as part of the new
self-init flow, loads the rest of the firmware images to
the device.

Store the image, so the ROM can load it to the device.

Signed-off-by: Golan Ben Ami <golan.ben.ami@intel.com>
automatic-review: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
iwl7000-tree: 1e5988aae7a0571203c889dbf9e655cd08566331