CHROMIUM: [media] s5p-mfc: Always signal source change event on format change
Currently the driver signals the source change event only in case of
a midstream resolution change, however the initial format detection
is also defined as a source change by the V4L2 codec API specification.
Fix this by signaling the event after the initial header is parsed as
well.

Also, remove the legacy behavior of VIDIOC_G_FMT, which used to
synchronously wait for hardware to finish parsing the header, because it
is against the V4L2 codec API specification and can actually deadlock
the userspace in case when an OUTPUT buffer that does not contain
headers is queued.

BUG=chromium:735783
TEST=video_decode_accelerator_unittest

Change-Id: I391c2ba28c2a14420332de3578e39511976d7e49
Signed-off-by: Tomasz Figa <tfiga@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/589087
Tested-by: Hirokazu Honda <hiroh@chromium.org>
Reviewed-by: Pawel Osciak <posciak@chromium.org>