CHROMIUM: tpm: ignore failed selftest in probe
If a tpm is in failed self-test mode, ContinueSelfTest
command results in TPM_FAILEDSELFTEST error, and probe
fails. The tpm device is not created, and that prevents
the OS from attempting any further recover operations
with the tpm. Instead, ignore the negative result of the
self test, and create the device - the chip is out there,
it's just in the failed state.

BUG=chromium:728130
TEST=put tpm in a (simulated) failed selftest mode, boot,
     check that /dev/tpm0 is present, and GetCapability
     commands succeed.

Change-Id: Ic3197e3385a893207f248c4a35cebf2bdbfe6bf3
Signed-off-by: Andrey Pronin <apronin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/596550
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>